Phase 6: documentation, validation, quality gates
Quality gates (all pass):
- mypy --strict: 0 errors (was 15 — fixed all type annotations)
- ruff check + format: 0 errors
- pytest: 55/55 pass

Documentation:
- README.md with physics overview, accuracy table, quick start, project structure
- CLAUDE.md updated with full API docs, physics implementation details
- All modules have complete docstrings

Validation:
- data/validate.py: full NASA comparison sweep across all 10 materials x
  5 energies x 3 angles with per-material RMS deviation reporting
- Supports --verbose, --json, --slabs flags

Code cleanup (code-simplifier + mypy agent):
- Extracted shared _resolve_material to server/api/__init__.py
- Fixed all mypy type annotations (float() casts, Any types, type: ignore)
- Removed dead _density_effect_correction function
- Updated .gitignore for validation artifacts

+++ config.py
@@ -163,7 +163,7 @@ def mean_ionization_potential_ev(z: int | float) -> float:
         raise ValueError(msg)
     if z_int in MEAN_IONIZATION_POTENTIALS_EV:
         return float(MEAN_IONIZATION_POTENTIALS_EV[z_int])
-    return 9.76 * z_int + 58.8 * z_int ** (-0.19)
+    return float(9.76 * z_int + 58.8 * z_int ** (-0.19))
 
 
 # ---------------------------------------------------------------------------
@@ -177,7 +177,7 @@ def thomas_fermi_radius_cm(z: int | float) -> float:
     if z <= 0:
         msg = f"Atomic number must be positive, got Z={z}"
         raise ValueError(msg)
-    return BOHR_RADIUS_CM * float(z) ** (-1.0 / 3.0)
+    return float(BOHR_RADIUS_CM * float(z) ** (-1.0 / 3.0))
 
 
 # ---------------------------------------------------------------------------
@@ -198,7 +198,7 @@ def backscatter_fraction(z: int | float, electron_energy_mev: float) -> float:
         msg = f"Electron energy must be positive, got {electron_energy_mev} MeV"
         raise ValueError(msg)
     z_f = float(z)
-    return min(0.005 * z_f**0.7 / (1.0 + 0.3 * electron_energy_mev), 0.5)
+    return float(min(0.005 * z_f**0.7 / (1.0 + 0.3 * electron_energy_mev), 0.5))
 
 
 # ---------------------------------------------------------------------------
